separate file to search and insert
